refactor: . Move taskSearchSuggestionText & taskSearchMetadataText new file src/ui/QuickSearchTasksModal.ts
